[shyun] 다양한 구체를 이용하여 삼각형 크기 줄이기, offset texture on/off, kd-tree offset realloc debug, camera move change

+++ Kd-treeConverter_v0.9/Kd-treeConverter/Kd-treeConstructor.h
@@ -94,8 +94,10 @@ extern TriangleList *g_pTriangleInfos;
 extern unsigned int  g_iTriangleSize;
 
 extern unsigned int  g_iKdTree_Level;
-extern unsigned int  g_iKdTree_TriOffset_Count;
-extern unsigned int  g_iKdTree_TriOffset_CountAlloc;
+//extern unsigned int  g_iKdTree_TriOffset_Count;
+//extern unsigned int  g_iKdTree_TriOffset_CountAlloc;
+extern unsigned long long  g_iKdTree_TriOffset_Count;
+extern unsigned long long  g_iKdTree_TriOffset_CountAlloc;
 extern unsigned int *g_pKdTree_TriOffset_Array;
 extern unsigned int  g_iKdTree_Node_Count;
 extern unsigned int  g_iKdTree_Node_CountAlloc;
